fix: retval type of rt_brk corrected to uintptr_t

+++ emodules/rt_base/rt_syscall.h
@@ -49,7 +49,7 @@
 // #define ERR_DRV_NOT_FND -111
 
 int rt_fstat(uintptr_t fd, uintptr_t sstat);
-int rt_brk(uintptr_t addr);
+uintptr_t rt_brk(uintptr_t addr);
 int rt_write(uintptr_t fd, char* content, size_t n);
 int rt_close(uintptr_t fd);
 int rt_gettimeofday(struct timeval* tv, struct timezone* tz);
